requirements.txt
captcha==0.2.4
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyrenn==0.1
